Fix few warnings.
While passing CString object works it is not a good practise. We do that
everywhere, but at least fix those places where compiler throws warning.

Handle POSITION mess, cast to integer type before casting to pointer.
Add __assume to let compiler know that values fits in the INT range.

+++ src/Subtitles/VobSubFileRipper.cpp
@@ -1033,7 +1033,7 @@ STDMETHODIMP CVobSubFileRipper::LoadParamFile(CString fn)
 
                         m_rd.selids[(BYTE)langnum] = true;
                     } else if (_istalpha(lang[0])) {
-                        n = _stscanf_s(lang, _T("%s"), langid, _countof(langid));
+                        n = _stscanf_s(lang, _T("%s"), langid, UINT(_countof(langid)));
                         if (n != 1) {
                             break;
                         }
